No limit for weather block length
Hi,

I suggest removing the test on the length of the block string. Depending on the format set by the user, valid strings can be rejected. Ex: a format that displays the name of a city that is more than 10 characters long (San Francisco), or a format that displays wind or other parameters.

Printing whatever is returned by the weather service is a good way to debug. Printing nothing can be confusing.