Properly use colors for rofication message count
Unlike the `battery` script, this script uses `{$color:-}` ahead of the
xrescat call. As `color` is always provided, it takes precedence and the
warning and critical colors are never used. This change ignores $color
and just uses the result from the respective xrescat calls.

Arguably the fallback values should not be #D8DEE9 for all, but I expect
the referenced resources to be there at all times.